+++ spec/support/auth_helper.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module AuthHelper
+  def http_login
+    request.env['HTTP_AUTHORIZATION'] =
+      ActionController::HttpAuthentication::Basic.encode_credentials('username', 'password')
+  end
+end